// packages/@ros-cdk/services/sag/lib/sag.generated.ts
export interface RosACLRuleProps {

    /**
     * @Property aclId: Access control ID.
     */
    readonly aclId: string;

    /**
     * @Property destCidr: Destination address, CIDR format and IP address range in IPv4 format.
     */
    readonly destCidr: string;

    /**
     * @Property destPortRange: Destination port range, 80/80.
     */
    readonly destPortRange: string;

    /**
     * @Property direction: Regular direction.
     * Value: in|out
     */
    readonly direction: string;

    /**
     * @Property ipProtocol: Protocol, not case sensitive.
     */
    readonly ipProtocol: string;

    /**
     * @Property policy: Access: accept|drop
     */
    readonly policy: string;

    /**
     * @Property sourceCidr: Source address, CIDR format and IP address range in IPv4 format.
     */
    readonly sourceCidr: string;

    /**
     * @Property sourcePortRange: Source port range, 80/80.
     */
    readonly sourcePortRange: string;

    /**
     * @Property description: Rule description information, ranging from 1 to 512 characters.
     */
    readonly description?: string;

    /**
     * @Property priority: Priority, ranging from 1 to 100.
     * Default: 1
     */
    readonly priority?: number;
}
// ...
export class RosACLRule extends ros.RosResource {
    /**
     * The resource type name for this resource class.
     */
    public static readonly ROS_RESOURCE_TYPE_NAME = "ALIYUN::SAG::ACLRule";

    /**
     * A factory method that creates a new instance of this class from an object
     * containing the properties of this ROS resource.
     */

    /**
     * @Attribute AcrId: Access control rule ID.
     */
    public readonly attrAcrId: any;

    public enableResourcePropertyConstraint: boolean;


    /**
     * @Property aclId: Access control ID.
     */
    public aclId: string;

    /**
     * @Property destCidr: Destination address, CIDR format and IP address range in IPv4 format.
     */
    public destCidr: string;

    /**
     * @Property destPortRange: Destination port range, 80/80.
     */
    public destPortRange: string;

    /**
     * @Property direction: Regular direction.
     * Value: in|out
     */
    public direction: string;

    /**
     * @Property ipProtocol: Protocol, not case sensitive.
     */
    public ipProtocol: string;

    /**
     * @Property policy: Access: accept|drop
     */
    public policy: string;

    /**
     * @Property sourceCidr: Source address, CIDR format and IP address range in IPv4 format.
     */
    public sourceCidr: string;

    /**
     * @Property sourcePortRange: Source port range, 80/80.
     */
    public sourcePortRange: string;

    /**
     * @Property description: Rule description information, ranging from 1 to 512 characters.
     */
    public description: string | undefined;

    /**
     * @Property priority: Priority, ranging from 1 to 100.
     * Default: 1
     */
    public priority: number | undefined;

    /**
     * Create a new `ALIYUN::SAG::ACLRule`.
     *
     * @param scope - scope in which this resource is defined
     * @param id    - scoped id of the resource
     * @param props - resource properties
     */
    constructor(scope: ros.Construct, id: string, props: RosACLRuleProps, enableResourcePropertyConstraint: boolean) {
        super(scope, id, { type: RosACLRule.ROS_RESOURCE_TYPE_NAME, properties: props });
        this.attrAcrId = ros.Token.asString(this.getAtt('AcrId'));

        this.enableResourcePropertyConstraint = enableResourcePropertyConstraint;
        this.aclId = props.aclId;
        this.destCidr = props.destCidr;
        this.destPortRange = props.destPortRange;
        this.direction = props.direction;
        this.ipProtocol = props.ipProtocol;
        this.policy = props.policy;
        this.sourceCidr = props.sourceCidr;
        this.sourcePortRange = props.sourcePortRange;
        this.description = props.description;
        this.priority = props.priority;
    }


    protected get rosProperties(): { [key: string]: any }  {
        return {
            aclId: this.aclId,
            destCidr: this.destCidr,
            destPortRange: this.destPortRange,
            direction: this.direction,
            ipProtocol: this.ipProtocol,
            policy: this.policy,
            sourceCidr: this.sourceCidr,
            sourcePortRange: this.sourcePortRange,
            description: this.description,
            priority: this.priority,
        };
    }
    protected renderProperties(props: {[key: string]: any}): { [key: string]: any }  {
        return rosACLRulePropsToRosTemplate(props, this.enableResourcePropertyConstraint);
    }
}